add: reorganized problems in easy-medium-hard set within blind75 package
