+++ Sorting/src/blind75/ListNode.java
@@ -4,18 +4,18 @@ package blind75;
  * Definition for singly-linked list node.
  * This is the standard ListNode structure used in LeetCode problems.
  */
-class ListNode {
-    int val;
-    ListNode next;
+public class ListNode {
+    public int val;
+    public ListNode next;
 
-    ListNode() {
+    public ListNode() {
     }
 
-    ListNode(int val) {
+    public ListNode(int val) {
         this.val = val;
     }
 
-    ListNode(int val, ListNode next) {
+    public ListNode(int val, ListNode next) {
         this.val = val;
         this.next = next;
     }
