+++ Sorting/src/blind75/easy/MissingNumber.java
@@ -1,4 +1,4 @@
-package blind75;
+package blind75.easy;
 
 import samples.Display;
