+++ Sorting/src/blind75/easy/ReverseLinkedList.java
@@ -1,5 +1,6 @@
-package blind75;
+package blind75.easy;
 
+import blind75.ListNode;
 import static blind75.LinkedListUtils.*;
 
 /**
