+++ Sorting/src/blind75/easy/TwoSum.java
@@ -1,4 +1,4 @@
-package blind75;
+package blind75.easy;
 
 import java.util.HashMap;
 import java.util.Map;
