+++ Sorting/src/blind75/medium/LongestSubstringWithoutRepeatingChars.java
@@ -1,4 +1,4 @@
-package blind75;
+package blind75.medium;
 
 import samples.Display;
