+++ Sorting/src/patterns/PatternTemplate.java
@@ -1,0 +1,57 @@
+package patterns;
+
+/**
+ * Pattern Documentation Template
+ *
+ * This package contains reusable patterns and techniques identified while solving
+ * LeetCode sorting problems. Each pattern file documents:
+ *
+ * 1. Pattern Description - What is this pattern?
+ * 2. When to Use - Problem characteristics that indicate this pattern
+ * 3. Implementation Template - Reusable code structure
+ * 4. Example Problems - Which problems use this pattern
+ * 5. Complexity Analysis - Time and space implications
+ * 6. Variations - Different forms of the pattern
+ * 7. Common Pitfalls - What to watch out for
+ *
+ * Patterns to be documented as problems are solved:
+ * - TwoPointerPattern.java
+ * - FrequencyCountPattern.java
+ * - CustomComparatorPattern.java
+ * - GreedyApproachPattern.java
+ * - BucketSortPattern.java
+ * - PartitionPattern.java
+ * - MergePattern.java
+ * - BinarySearchPattern.java
+ * - InPlaceTransformPattern.java
+ * - MathOptimizationPattern.java
+ *
+ * Each pattern will be extracted as common solutions emerge across problems.
+ */
+public class PatternTemplate {
+
+    /**
+     * Pattern Name: [To be defined]
+     *
+     * Description:
+     * [Detailed explanation of what this pattern does]
+     *
+     * When to Use:
+     * - [Condition 1]
+     * - [Condition 2]
+     * - [Condition 3]
+     *
+     * Time Complexity: O(?)
+     * Space Complexity: O(?)
+     *
+     * Example Problems:
+     * - Problem #X: [Problem Name]
+     * - Problem #Y: [Problem Name]
+     */
+
+    public static void main(String[] args) {
+        System.out.println("Patterns package - Documentation and reusable techniques");
+        System.out.println("Patterns will be documented as problems are solved.");
+    }
+}
+
